+++ inst/templates/main_script.R
@@ -136,7 +136,7 @@ reportPDF(fGroups, path = "report", reportFGroups = TRUE, formConsensus = {{ if 
 
 {{ endCodeBlock() }}
 {{ optionalCodeBlock("MD" %in% reportFormats) }}
-reportMD(fGroups, path = "report", reportFGroups = TRUE, reportChord = TRUE, formConsensus = {{ if (formulaOpts$algo != "") "formulas" else "NULL" }}, reportFormulaSpectra = TRUE,
+reportMD(fGroups, path = "report", reportPlots = c("chord", "venn", "upset", "eics", "formulas"), formConsensus = {{ if (formulaOpts$algo != "") "formulas" else "NULL" }},
          compounds = {{ if (identOpts$algo != "") "compounds" else "NULL" }}, compoundNormalizeScores = TRUE,
          components = {{ if (componentOpts$algo != "") "components" else "NULL" }}, MSPeakLists = {{ if (formulaOpts$algo != "" || identOpts$algo != "") "plists" else "NULL" }},
          selfContained = TRUE)