+++ tests/testthat/test-compounds_clust.R
@@ -70,7 +70,7 @@ test_that("reporting works", {
     for (grp in names(clusters(compsClust)))
         checkmate::expect_file_exists(getWorkPath("compounds", sprintf("%s-%s-clusters.pdf", class(fGroups), grp))) # UNDONE: check col name
     
-    expect_file(reportMD(fGroups, getWorkPath(), reportChord = FALSE, reportFGroups = FALSE,
+    expect_file(reportMD(fGroups, getWorkPath(), reportPlots = "none",
                          compounds = compounds, MSPeakLists = plists, compsCluster = compsClust),
                 getWorkPath("report.html"))
 })
@@ -80,6 +80,6 @@ test_that("reporting empty object works", {
     
     expect_error(reportCSV(fGroups, getWorkPath(), compsCluster = compsClustEmpty), NA)
     expect_error(reportPDF(fGroups, getWorkPath(), compsCluster = compsClustEmpty), NA)
-    expect_file(reportMD(fGroups, getWorkPath(), reportChord = FALSE, reportFGroups = FALSE,
+    expect_file(reportMD(fGroups, getWorkPath(), reportPlots = "none",
                          compsCluster = compsClustEmpty), getWorkPath("report.html"))
 })